+++ CHANGELOG.md
@@ -33,6 +33,7 @@ Uliweb Change Log
 * add None patch process, you can set '', 'empty, 'exception'.
 * move uliweb/orm/middle*.py to uliweb/contrib/orm
 * move uliweb/i18n/middle_i18n.py to uliweb/contrib/i18n
+* move storage from core to utils directory
 
 0.2.6 Version
 -----------------